+++ src/pubsub.ts
@@ -5,7 +5,9 @@ export interface PubSubOptions {
   eventEmitter?: EventEmitter;
 }
 
-export class PubSub extends PubSubEngine {
+export class PubSub<
+  Events extends { [event: string]: unknown } = Record<string, never>
+> extends PubSubEngine {
   protected ee: EventEmitter;
   private subscriptions: { [key: string]: [string, (...args: any[]) => void] };
   private subIdCounter: number;
@@ -17,12 +19,18 @@ export class PubSub extends PubSubEngine {
     this.subIdCounter = 0;
   }
 
-  public publish(triggerName: string, payload: any): Promise<void> {
+  public publish<K extends keyof Events>(
+    triggerName: K & string,
+    payload: Events[K] extends never ? any : Events[K]
+  ): Promise<void> {
     this.ee.emit(triggerName, payload);
     return Promise.resolve();
   }
 
-  public subscribe(triggerName: string, onMessage: (...args: any[]) => void): Promise<number> {
+  public subscribe<K extends keyof Events>(
+    triggerName: K & string,
+    onMessage: (...args: any[]) => void
+  ): Promise<number> {
     this.ee.addListener(triggerName, onMessage);
     this.subIdCounter = this.subIdCounter + 1;
     this.subscriptions[this.subIdCounter] = [triggerName, onMessage];